FIX: remove unnecesasry stubbing and move stub verification down into base class
