+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/config/MessagingMultiNodeTestBase.kt
@@ -1,21 +1,11 @@
 package uk.gov.justice.digital.hmpps.personrecord.config
 
-import com.github.tomakehurst.wiremock.client.WireMock.deleteRequestedFor
-import com.github.tomakehurst.wiremock.client.WireMock.getRequestedFor
-import com.github.tomakehurst.wiremock.client.WireMock.postRequestedFor
-import com.github.tomakehurst.wiremock.client.WireMock.putRequestedFor
-import com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo
-import com.github.tomakehurst.wiremock.client.WireMock.urlMatching
-import com.github.tomakehurst.wiremock.http.RequestMethod
 import com.github.tomakehurst.wiremock.stubbing.Scenario.STARTED
-import org.assertj.core.api.Assertions.fail
 import org.awaitility.kotlin.await
 import org.awaitility.kotlin.matches
 import org.awaitility.kotlin.untilCallTo
-import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.extension.ExtendWith
-import org.mockito.kotlin.verify
 import org.springframework.beans.factory.annotation.Autowired
 import software.amazon.awssdk.services.sns.model.MessageAttributeValue
 import software.amazon.awssdk.services.sqs.model.PurgeQueueRequest
@@ -255,25 +245,6 @@ abstract class MessagingMultiNodeTestBase : IntegrationTestBase() {
     currentScenarioState: String? = STARTED,
     nextScenarioState: String? = STARTED,
   ) = stubGetRequest(scenarioName, currentScenarioState, nextScenarioState, "/prisoner/${apiResponseSetup.prisonNumber}", prisonerSearchResponse(apiResponseSetup))
-
-  @AfterEach
-  fun after() {
-    wiremock.stubMappings.forEach {
-      when (it.request.method) {
-        RequestMethod.GET -> {
-          if (it.request.url != null) {
-            wiremock.verify(getRequestedFor(urlEqualTo(it.request.url)))
-          } else {
-            wiremock.verify(getRequestedFor(urlMatching(it.request.urlPathPattern)))
-          }
-        }
-        RequestMethod.POST -> wiremock.verify(postRequestedFor(urlEqualTo(it.request.url)))
-        RequestMethod.DELETE -> wiremock.verify(deleteRequestedFor(urlEqualTo(it.request.url)))
-        RequestMethod.PUT -> wiremock.verify(putRequestedFor(urlEqualTo(it.request.url)))
-        else -> fail()
-      }
-    }
-  }
 
   @BeforeEach
   fun beforeEachMessagingTest() {
