+++ src/test/kotlin/uk/gov/justice/digital/hmpps/personrecord/service/SearchServiceIntTest.kt
@@ -115,7 +115,6 @@ class SearchServiceIntTest : IntegrationTestBase() {
       ),
     )
 
-    stubOneHighConfidenceMatch()
     stubPersonMatchScores(person.matchId)
     val candidateRecords = searchService.findCandidateRecordsWithUuid(person)
